+++ CHANGES.md
@@ -86,11 +86,13 @@ session context for the listener's invocation.
 * `CipherInformation#getBlockSize` has been renamed to `getKdfSize` in order to emphasize that its value
 represents the number of bytes used to derive the cipher's secret key value and not the cipher's underlying
 block size.
-    
+
     * See [SSHD-873](https://issues.apache.org/jira/browse/SSHD-873) enhancement remark below
     for the cipher block size information.
 
 ## Behavioral changes and enhancements
+
+* [SSHD-708](https://issues.apache.org/jira/browse/SSHD-708) - Add support for password encrypted OpenSSH private key files
 
 * [SSHD-757](https://issues.apache.org/jira/browse/SSHD-757) - Added hooks and some initial code to allow (limited) usage
 of [OpenPGP](https://www.openpgp.org/) key files - e.g. in `authorized_keys` files or as client identities.